feat(types): enhance shared types package with new definitions and updates
- Added `@vernisai/types` dependency to API and web packages for centralized type usage
- Migrated existing dashboard types to the shared `@vernisai/types` package
- Introduced `ConversationSummary` type for improved dashboard data handling
- Updated imports in various components to utilize the new shared types
- Refactored `turbo.json` to align with Turbo 2.0 changes
- Validated changes through successful build and testing processes
