+++ apps/web/package.json
@@ -17,6 +17,7 @@
     "@trpc/react-query": "^10.37.1",
     "@trpc/server": "^10.37.1",
     "@vernisai/api": "*",
+    "@vernisai/types": "^0.1.0",
     "@vernisai/ui": "^0.0.0",
     "lucide-react": "^0.335.0",
     "react": "^18.3.1",
