+++ changes/2025-03-25.md
@@ -79,3 +79,11 @@
 - Designed for future database schema integration
 - Improved separation of concerns between packages
 - Eliminated type duplication between client and server
+
+## Migrated to shared @vernisai/types - 02:14 AM
+
+- Migrated dashboard types from API and web packages to shared @vernisai/types package
+- Created ConversationSummary type for dashboard views
+- Added TestMessage type to chat/message.ts
+- Fixed imports in web and API packages to use the shared types
+- Updated turbo.json to use "tasks" instead of "pipeline" for Turbo 2.0 compatibility
