+++ packages/types/src/dashboard/conversation.ts
@@ -1,0 +1,14 @@
+/**
+ * Conversation types for the dashboard
+ */
+
+/**
+ * Represents a conversation summary for dashboard
+ */
+export type ConversationSummary = {
+  id: string;
+  title: string;
+  lastMessageAt: Date;
+  model: string;
+  snippet: string;
+};
