+++ packages/types/src/index.ts
@@ -15,3 +15,9 @@ export * from "./user";
 
 // Export common utility types
 export * from "./common";
+
+// Re-export specific types to ensure backward compatibility
+// as we split types into proper domains
+export type { TestMessage } from "./chat/message";
+export type { AgentTemplate } from "./dashboard/agent";
+export type { ConversationSummary } from "./dashboard/conversation";
